requirements.txt
fastapi[all]
uvicorn
python-dotenv
pydantic